misc/HtmlTestSuite/results/test-0576.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Shapes Module Level 1: getComputedStyle().shapeMargin</title>
<link href="https://drafts.csswg.org/css-shapes/#shape-margin-property" rel="help"/>
<meta content="shape-margin computed value is non-negative &lt;length-percentage&gt;." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/computed-testcommon.js"/>
</head>
<body>
<style>
  #target {
    font-size: 40px;
  }
</style>
<div id="target"/>
<script>
test_computed_value("shape-margin", "calc(10px + 0.5em)", "30px");
test_computed_value("shape-margin", "calc(10px - 0.5em)", "0px");
test_computed_value("shape-margin", "50%");
</script>
</body>
</html>
